requirements.txt
Pillow==11.0.0
PySide6
selenium==4.27.1
